WebSocket message handling

+++ src/main/java/pki/sqlTables/content/authors/AuthorRepository.java
@@ -3,6 +3,6 @@ package pki.sqlTables.content.authors;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface AuthorRepository extends CrudRepository<AuthorEntity, Integer> {
+public interface AuthorRepository extends CrudRepository<AuthorEntity, Long> {
 
 }
