+++ src/main/java/pki/sqlTables/content/books/BookRepository.java
@@ -2,5 +2,5 @@ package pki.sqlTables.content.books;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface BookRepository extends CrudRepository<BookEntity, Integer> {
+public interface BookRepository extends CrudRepository<BookEntity, Long> {
 }
